backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.0
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jaconv==0.4.0
Janome==0.5.0
japanize-matplotlib==1.1.3
Jinja2==3.1.5
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mecab-python3==1.0.10
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
selenium==3.141.0
setuptools==75.4.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.38
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
unidic-lite==1.0.8
urllib3==1.26.16
webdriver-manager==4.0.2
Werkzeug==3.1.3
wheel==0.45.0
wordcloud==1.9.4
wrapt==1.16.0
psycopg2-binary